# src/docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "Это API для управления пользователями, товарами и поставщиками.",
        "title": "API документация",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/product/create": {
            "post": {
                "description": "Добавляет новый товар в систему",
                "tags": [
                    "products"
                ],
                "summary": "Создать товар",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/response.CreateProduct"
                            }
                        }
                    },
                    "description": "Данные товара",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка при разборе данных",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка при создании товара",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/product/delete/{id}": {
            "delete": {
                "description": "Удаляет товар по его UUID",
                "tags": [
                    "products"
                ],
                "summary": "Удалить товар",
                "parameters": [
                    {
                        "description": "UUID товара",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный формат UUID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ошибка при удалении товара",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/product/productList": {
            "get": {
                "description": "Возвращает список всех товаров",
                "tags": [
                    "products"
                ],
                "summary": "Получить список товаров",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/response.ProductResponse"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ошибка при получении товаров",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/product/updateQuantity": {
            "patch": {
                "description": "Уменьшает количество указанного товара на складе",
                "tags": [
                    "products"
                ],
                "summary": "Уменьшить количество товара на складе",
                "parameters": [
                    {
                        "description": "UUID товара",
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Количество для уменьшения",
                        "name": "quantity",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный формат UUID или количества",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ошибка при уменьшении товара",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/product/{id}": {
            "get": {
                "description": "Возвращает информацию о товаре по его UUID",
                "tags": [
                    "products"
                ],
                "summary": "Получить товар по ID",
                "parameters": [
                    {
                        "description": "UUID товара",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ProductResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный формат ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ошибка при получении товара",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "//localhost:5000/api/v1"
        }
    ],
    "components": {
        "schemas": {
            "response.CreateProduct": {
                "type": "object",
                "properties": {
                    "available_stock": {
                        "type": "integer"
                    },
                    "category": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "supplierID": {
                        "type": "string"
                    }
                }
            },
...
            "response.ProductResponse": {
                "type": "object",
                "properties": {
                    "ID": {
                        "type": "string"
                    },
                    "available_stock": {
                        "type": "integer"
                    },
                    "category": {
                        "type": "string"
                    },
                    "imageID": {
                        "type": "string"
                    },
                    "lastUpdateDate": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "supplierID": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}